backend/requirements.txt
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jira==3.13.0
groq==0.4.1
reportlab==4.0.7
python-docx==0.8.11
aiofiles==23.2.1
